ref(profiling): Refactor profile processing into its own crate
