+++ relay-server/src/utils/mod.rs
@@ -4,7 +4,6 @@ mod dynamic_sampling;
 mod error_boundary;
 mod multipart;
 mod param_parser;
-mod profile;
 mod rate_limits;
 mod request;
 mod shutdown;
@@ -26,7 +25,6 @@ pub use self::dynamic_sampling::*;
 pub use self::error_boundary::*;
 pub use self::multipart::*;
 pub use self::param_parser::*;
-pub use self::profile::*;
 pub use self::rate_limits::*;
 pub use self::request::*;
 pub use self::shutdown::*;
